Revise the instructions for building a visualizer
Revise the p5 visualizer template and Differences visualizer to match.
