+++ src/visualizers-workbench/visualizers.ts
@@ -8,7 +8,6 @@ const vizMODULES: {[key: string]: VisualizerExportModule} = {}
 
 // for each file in the list of visualizer files...
 if (import.meta.env.VITE_WORKBENCH === '1') {
-    console.log('Importing workbench visualizers')
     for (const file in vizFiles) {
         const mod = vizFiles[file] as {exportModule: VisualizerExportModule}
         // if it has an export module...
